+++ smirnov.alexandr/S2/main.cpp
@@ -1,7 +1,8 @@
 #include <iostream>
+#include <list>
+#include <iomanip>
 #include "read_expressions.hpp"
 #include "conversion.hpp"
-#include <list>
 
 int main(int argc, char * argv[])
 {
@@ -12,7 +13,7 @@ int main(int argc, char * argv[])
   }
 
   std::list< std::string > expressions = smirnov::readExpressions(filename);
-  std::list< double > results;
+  std::list< long long > results;
 
   std::list< std::string >::iterator it;
   for (it = expressions.begin(); it != expressions.end(); ++it)
@@ -21,7 +22,7 @@ int main(int argc, char * argv[])
     try
     {
       smirnov::Queue< std::string > postfix = smirnov::infixToPostfix(expr);
-      double result = evaluatePostfix(postfix);
+      long long result = evaluatePostfix(postfix);
       results.push_front(result);
     }
     catch (const std::exception & e)
@@ -32,7 +33,7 @@ int main(int argc, char * argv[])
   }
   if (!results.empty())
   {
-    std::list< double >::iterator rit = results.begin();
+    std::list< long long >::iterator rit = results.begin();
     std::cout << *rit;
     ++rit;
     for (; rit != results.end(); ++rit)
